+++ docs/index.rst
@@ -160,7 +160,7 @@ As of xphyle 2.0.0, you can easily open subprocesses using the ``xphyle.popen`` 
     from xphyle.utils import exec_process
     exec_process('cat', 'foo', stdout='myfile.gz')
 
-In addition, ``open_`` and ``xopen`` can open subprocesses. The primariy difference is that ``popen`` enables customization of stdin, stdout, and stderr, whereas opening a process through ``open_`` or ``xopen`` uses default behavior of opening PIPEs for all of the streams, and wrapping the PIPE indicated by the file mode. For example::
+In addition, ``open_`` and ``xopen`` can open subprocesses. The primary difference is that ``popen`` enables customization of stdin, stdout, and stderr, whereas opening a process through ``open_`` or ``xopen`` uses default behavior of opening PIPEs for all of the streams, and wrapping the PIPE indicated by the file mode. For example::
     
     # write to the process stdin
     with open_('|cat', 'wt') as proc:
@@ -181,7 +181,7 @@ Note that with ``open_`` and ``xopen``, the system command must be specified as 
 Buffers
 ~~~~~~~
 
-As of xphyle 2.1.0, ``open_`` and ``xopen`` can also open buffer types. A buffer is an instance of ``io.StringIO`` or ``io.BytesIO`` (or similar) -- basically an in memory read/write buffer. Passing open buffer objects worked before (they were treated as file-like), but now there is a special file type -- ``FileType.BUFFER`` -- that will cause them to be handeled  a bit differently. In addition, you can now pass ``str`` or ``bytes`` (the type objects) to automatically create the corresponding buffer type::
+As of xphyle 2.1.0, ``open_`` and ``xopen`` can also open buffer types. A buffer is an instance of ``io.StringIO`` or ``io.BytesIO`` (or similar) -- basically an in memory read/write buffer. Passing open buffer objects worked before (they were treated as file-like), but now there is a special file type -- ``FileType.BUFFER`` -- that will cause them to be handled  a bit differently. In addition, you can now pass ``str`` or ``bytes`` (the type objects) to automatically create the corresponding buffer type::
 
     with open_(str) as buf:
         buf.write('foo')
@@ -191,6 +191,11 @@ As of xphyle 2.1.0, ``open_`` and ``xopen`` can also open buffer types. A buffer
     with open_(bytes, compression='gzip') as buf:
         buf.write('foo')
     compressed_foo = buf.getvalue()
+
+You can also create readable buffers by passing the string/bytes to read instead of a path, and explicitly specifying the file type::
+
+    with open_("This is a string I want to read", file_type=FileType.BUFFER) as buf:
+        buf_str = buf.read()
 
 Reading/writing data
 ~~~~~~~~~~~~~~~~~~~~
@@ -220,30 +225,37 @@ There are pairs of methods for reading/writing text and binary data using iterat
 
 There's another pair of methods for reading/writing key=value files::
     
+    from collections import OrderedDict
     from xphyle.utils import read_dict, write_dict
-    cats = dict(fluffy='calico', droopy='tabby', sneezy='siamese')
+    cats = OrderedDict((fluffy,'calico'), (droopy,'tabby'), (sneezy,'siamese'))
     write_dict(cats, 'cats.txt.gz')
-    # change from '=' to '\t' delimited
+    # change from '=' to '\t' delimited; preserve the order of the items
     write_dict(
-        read_dict(cats, 'cats.txt.gz'),
+        read_dict(cats, 'cats.txt.gz', ordered=True),
         'cats.tsv', sep='\t')
 
 You can also read from delimited files such as csv and tsv::
     
     from xphyle.utils import read_delimited, read_delimited_as_dict
+    
     class Dog(object):
[…]
