### 2017-05-03 #253 @pjotrp
@peterjc can you suggest a reviewer? I think this is an interesting addition to python bioinformatics (and beyond)
