### 2017-05-03 #253 @peterjc
Hmm. Reviewing a Python library for efficient access to compressed files - I'd think someone doing IO heavy Python/C would be best. @ctb or @mr-c could you suggest someone?
