### 2017-05-03 816eb8d
hits -> hints
