### 2017-05-04 #255 @whedon
[REVIEW]: xphyle: Extraordinarily simple file handling
**Submitting author:** @jdidion (<a href="http://orcid.org/0000-0002-8111-6261">John Didion</a>)
**Repository:** <a href="https://github.com/jdidion/xphyle" target ="_blank">https://github.com/jdidion/xphyle</a>
**Version:** 3.0.1
