### 2017-05-04 #255 @pjotrp
@jdidion to make life a bit easier for us, can you go through above check list and see if you match the important criteria? Thanks! After confirmation we'll start review.
