### 2017-05-04 037ad2c
Update license to be OSI compliant
I created the xphyle project as a private citizen. 100% of the code
contributed to the project was by myself as a US Government employee.
Thus, the project itself is copyright under an MIT license, but all of
the code within the project is granted without any reservation of
copyright.
