+++ setup.py
@@ -20,7 +20,7 @@ setup(
     url='https://github.com/jdidion/xphyle',
     author='John Didion',
     author_email='john.didion@nih.gov',
-    license='Public Domain',
+    license='MIT',
     packages = ['xphyle'],
     install_requires = requirements,
     tests_require = ['pytest', 'pytest-cov'],
@@ -28,6 +28,7 @@ setup(
         'Development Status :: 5 - Production/Stable',
         'Intended Audience :: Developers',
         'Topic :: Software Development :: Libraries :: Python Modules',
+        'License :: MIT',
         'License :: Public Domain',
         'Programming Language :: Python :: 3.3',
         'Programming Language :: Python :: 3.4',
