### 2017-05-04 #255 @jdidion
@pjotrp Yes, it meets all the guidelines.
