### 2017-05-05 #255 @pjotrp
@ctb please review at your convenience :)
