### 2017-05-22 jdidion/xphyle#9 @ctb
tests import error on python 3.5.1, stock install/venv
% python setup.py test

```
...
  File "/Users/t/dev/xphyle/tests/test_formats.py", line 4, in <module>
    from xphyle.formats import *
  File "/Users/t/dev/xphyle/xphyle/__init__.py", line 13, in <module>
    from xphyle.formats import FORMATS, THREADS
  File "/Users/t/dev/xphyle/xphyle/formats.py", line 12, in <module>
    from xphyle.paths import (
  File "/Users/t/dev/xphyle/xphyle/paths.py", line 13, in <module>
    from xphyle.types import (
  File "/Users/t/dev/xphyle/xphyle/types.py", line 346, in <module>
    AnyChar = Union[bytes, Text]
NameError: name 'Text' is not defined
```
