### 2017-05-22 jdidion/xphyle#9 @ctb
This turns out to cause problems with a basic import, `from xphyle import *`. Suggestions on how to proceed are welcome :). Otherwise I'll spend some time googling about.
