### 2017-05-22 jdidion/xphyle#9 @jdidion
Oof - so this change was introduced in 3.5.2. Unfortunately there's no mention of that in the docs, so I assumed that it had been around since 3.5. I will fix this by moving to the new 'typing' module (https://pypi.org/project/typing/) and adding it as a dependency for python < 3.6.
