### 2017-05-22 89c6670
Force use of backports.typing in python < 3.6
