### 2017-05-22 8f11dc6
Fix license name
