### 2017-05-22 jdidion/xphyle#9 @jdidion
Fixed in 3.0.2. The fix for this isn't pretty, but it works: I have to force the use of backports.typing for versions prior to 3.6.

On a side note: It's looking like 3.6 is going to be regarded as the first feature-stable 3.x version, so I'm hoping most people on 3.x will upgrade at some point in the near future. I'll probably drop support for pre-3.6 at the next major release (4.0) - which probably won't be for a while, as I consider the xphyle API to be stable as of 3.0.
