### 2017-05-23 #255 @jdidion
Is it still not working for you after upgrading to 3.0.2?
