### 2017-06-14 jdidion/xphyle#11 @ctb
Small review issues
This is a running issue for items associated with [my review for JOSS](https://github.com/openjournals/joss-reviews/issues/255).

- [x] a pointer to the github issue tracker should probably be added in the README, as obvious as it is...
- [x] an explicit pointer to the stable docs on readthedocs could be added (https://xphyle.readthedocs.io/en/stable/) - the little 'docs' button is not always so clear and in any case points to latest/ which might not be a release.
- [x] maybe spellcheck the docs? I found 'miscelanenous' while reading.
- [x] maybe have 'make test' run 'python -m pytest' for those of us using too many virtualenvs and with lazy install habits. (Not a big deal, obviously.)
