### 2017-06-14 #255 @ctb
Very nice - works well for me and I've already adopted the `xopen` idiom in my own projects! I'm sad that it's Python 3 and above only, as otherwise I'd be able to use it in existing projects.

I left a small pile of issues [here](https://github.com/jdidion/xphyle/issues/11) but nothing particularly problematic.

The only thing I'm stuck with this whether and how to evaluate performance claims. @jdidion any suggestions on something to benchmark?
