### 2017-06-14 #255 @jdidion
Thanks for the speedy review and the kind words! The Python 3 only decision is definitely a selfish one - there are big i/o differences from 2 to 3 that I don’t really want to hassle with. I’m happy to accept contributions from anyone that wants to take on the back port.

Regarding the benchmark - this should be straight-forward. I’ll just dummy up a couple of files (lorem ipsum, fastq), compress them, and then decompress with system-level decompression on versus off.

I’ll take care of this and the rest of the issues this week.

Thanks again,

John

> On Jun 13, 2017, at 8:05 PM, C. Titus Brown <notifications@github.com> wrote:
>
> Very nice - works well for me and I've already adopted the xopen idiom in my own projects! I'm sad that it's Python 3 and above only, as otherwise I'd be able to use it in existing projects.
>
> I left a small pile of issues here <https://github.com/jdidion/xphyle/issues/11> but nothing particularly problematic.
>
> The only thing I'm stuck with this whether and how to evaluate performance claims. @jdidion <https://github.com/jdidion> any suggestions on something to benchmark?
>
