### 2017-06-14 jdidion/xphyle#11 @jdidion
Hmm..mock is part of the unittest package starting in python 3.3. I am testing against all versions of python 3.3-3.7 with travis and it works fine without including mock as a dependency.

> On Jun 13, 2017, at 8:02 PM, C. Titus Brown <notifications@github.com> wrote:
>
> This is a running issue for items associated with my review for JOSS <https://github.com/openjournals/joss-reviews/issues/255>.
>
>  a pointer to the github issue tracker should probably be added in the README, as obvious as it is...
>  an explicit pointer to the stable docs on readthedocs could be added (https://xphyle.readthedocs.io/en/stable/ <https://xphyle.readthedocs.io/en/stable/>) - the little 'docs' button is not always so clear and in any case points to latest/ which might not be a release.
>  maybe spellcheck the docs? I found 'miscelanenous' while reading.
>  the tests depend on 'mock', not just 'pytest-cov'.
>  maybe have 'make test' run 'python -m pytest' for those of us using too many virtualenvs and with lazy install habits. (Not a big deal, obviously.)
