### 2017-06-14 jdidion/xphyle#11 @ctb
yep, that was a holdover from a silly problem I had w/versions. I already updated the issue but you must have been working off the e-mail :)
