### 2017-06-14 f629a50
add basic performance testing
