+++ setup.py
@@ -23,7 +23,10 @@ setup(
     license='MIT',
     packages = ['xphyle'],
     install_requires = requirements,
-    tests_require = ['pytest', 'pytest-cov'],
+    extras_require = {
+        'performance' : ['lorem']
+    },
+    tests_require = ['pytest', 'pytest-cov', ],
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Intended Audience :: Developers',
