### 2017-06-14 6af0674
remove unneeded tqdm package
