### 2017-06-14 #255 @jdidion
Added basic performance testing. I generate two sets of files - one using lorem ipsum, and one by randomly generating sequences in fastq format. I then three separate operations: reading the lines of the file using the python gzip package directly (use_system=None), using xopen with use_system option==False, and with use_system==True. Things run 4-5x faster using system-level compression (on my 2013 Macbook Pro).

Lorem ipsum:
    Timing of 10 Lorem Ipsum tests with total size 232,005,276 characters and
    use_system = None: 0.0026 sec

    Timing of 10 Lorem Ipsum tests with total size 232,005,276 characters and
    use_system = True: 0.0005 sec

    Timing of 10 Lorem Ipsum tests with total size 232,005,276 characters and
    use_system = False: 0.0027 sec

Fastq
    Timing of 10 fastq tests with total size 672,053,460 characters and
    use_system = None: 0.0293 sec

    Timing of 10 fastq tests with total size 672,053,460 characters and
    use_system = True: 0.0065 sec

    Timing of 10 fastq tests with total size 672,053,460 characters and
    use_system = False: 0.0202 sec
