### 2017-06-14 #255 @jdidion
Oops... There was a mistaken divide by 1000, so multiply all those timings by 1000
