### 2017-06-14 2cde9f1
No divide by 1000 necessary - duration is already in seconds
