### 2017-06-14 96f46a1
Only keep 2 sig digits
