### 2017-06-14 06e53d2
Spell check docs, fix other small review issues
