### 2017-06-14 57cfa73
add ability to specify additional pytest options
