### 2017-06-14 782bf28
Fix #12: don't rely on empty string to signify a resolved but unavailable executable path
