+++ CHANGES.md
@@ -1,7 +1,8 @@
-v3.0.2 (dev)
-------------
+v3.0.2 (2017.06.14)
+-------------------
 * Forcing use of backports.typing for python < 3.6.
 * Added basic performance testing.
+* Fixed #12: xphyle not recognizing when system-level lzma not installed.
 
 v3.0.1 (2017.04.29)
 -------------------
