### 2017-06-14 38010a7
skip compression tests if system executable not available
