### 2017-06-17 #255 @ctb
works well, thanks! but pre-Python 3.6, 'random.choices' doesn't exist.

You might consider putting in CI for the earliest version of Python you
want to support ;)
