### 2017-06-17 #255 @jdidion
Ah right, thanks. I couldn't find any backports for the random package, so I just copied and adapted the function from the python source (with attribution of course).
